Decreased delays
- Decreased file check delay

Important:
I was not able to successfully fully host the Think-cell Server hosting program on Azure's VM for very long. Unfortunately the free credits on Azure were limited and I did not want to spend money on something that was not being utilized by anyone just for the sake of it.

But as far as everything goes, this project is finished.